requirements-mac.txt
# Core Data Science Libraries
pandas>=2.0.0,<2.3.0
numpy>=1.24.0,<2.0.0

# Visualization
matplotlib>=3.7.0,<3.10.0
seaborn>=0.12.0,<0.14.0
plotly>=5.14.0,<5.25.0
kaleido>=0.2.1

# Machine Learning (for Phase 3)
scikit-learn>=1.3.0,<1.6.0
scipy>=1.10.0,<1.15.0

# Spotify API
spotipy>=2.23.0

# Streamlit for Dashboard
streamlit>=1.28.0,<1.40.0

# Additional utilities
python-dotenv>=1.0.0
tenacity>=8.2.0  # Retry logic with exponential backoff

# Cloud Storage (Cloudflare R2 / S3-compatible)
boto3>=1.28.0
pyarrow>=14.0.0  # For Parquet file format

# Jupyter
jupyter>=1.0.0
ipython>=8.12.0
notebook>=7.0.0

# Optional: XGBoost for Phase 3 (comment out if issues)
# xgboost>=1.7.0
